+++ Shared/utils/pcb.c
@@ -4,7 +4,7 @@ Pcb *pcb_create(
    uint32_t pid,
    uint32_t tamano,
    Tabla_paginas *tabla_paginas,
-   double estimado_rafaga,
+   uint32_t estimado_rafaga,
    t_list* instrucciones){
        Pcb *pcb = (Pcb *) malloc(sizeof(Pcb));
        pcb->pid = pid;
