+++ Shared/utils/pcb.h
@@ -12,7 +12,7 @@ typedef struct {
    uint32_t tamano;
    uint32_t program_counter;
    Tabla_paginas *tabla_paginas;
-   double estimado_rafaga;
+   uint32_t estimado_rafaga;
    t_list* instrucciones;
 } Pcb;
 
@@ -20,7 +20,7 @@ Pcb *pcb_create(
    uint32_t pid,
    uint32_t tamano,
    Tabla_paginas *tabla_paginas,
-   double estimado_rafaga,
+   uint32_t estimado_rafaga,
    t_list* instrucciones);
 
 t_paquete *pcb_serializar(Pcb *pcb);
